Update logger setup and import statements in constants.dart
Added import for 'package:flutter/foundation.dart' for enabling different logger levels depending on the debug mode. Updated the logger setup to include PrettyPrinter and to set the level to verbose during debug mode and nothing otherwise. This allows for more detailed logging during development and silent logging in production.