Bump version and update password visibility icon
This commit bumps the shared_utils package version to 0.1.0-alpha.20, as reflected in the pubspec.yaml and CHANGELOG.md files. It also reverses the eye icon status in the password field widget for a more intuitive understanding of the password visibility status: now the 'eye' icon means the password is visible and 'eye_off' means it is hidden.