Commit message: Bump version and remove widget disposal
This commit increments the shared_utils package version to 0.1.0-alpha.21 in the pubspec.yaml and CHANGELOG.md files. Additionally, it removes the disposal method from the FilledTextField widget to avoid the potential issue of prematurely disposing a still-needed controller. Changes are included in the lib/src/widgets/app.text.field.dart file.